feat: refactor dashboard components for improved structure and performance, integrate new data sources
