--- components/calendar-01.tsx
@@ -1,21 +1,0 @@
-"use client"
-
-import * as React from "react"
-
-import { Calendar } from "@/components/ui/calendar"
-
-export default function Calendar01() {
-  const [date, setDate] = React.useState<Date | undefined>(
-    new Date(2025, 5, 12)
-  )
-
-  return (
-    <Calendar
-      mode="single"
-      defaultMonth={date}
-      selected={date}
-      onSelect={setDate}
-      className="rounded-lg border shadow-sm"
-    />
-  )
-}
