+++ types/campaign.ts
@@ -39,3 +39,21 @@ export enum statusCampaign {
   paused = "paused",
   completed = "completed",
 }
+
+
+export interface ChartData {
+  date: string;
+  sent: number;
+  opened: number;
+  replied: number;
+  bounced: number;
+  clicked: number;
+  formattedDate: string;
+}
+
+export interface MetricToggle {
+  key: keyof ChartData;
+  label: string;
+  color: string;
+  visible: boolean;
+}
